login function

+++ app/Http/Controllers/TasksController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 
-Use App\Task;
+use App\Tasks;
 
 class TasksController extends Controller
 {
@@ -15,14 +15,25 @@ class TasksController extends Controller
      */
     public function index()
     {
-        //
-        $tasks = Task::paginate(25);
-        
-        return view('tasks.index', [
-            'tasks'=>$tasks,
-            ]);
-        
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザの投稿も取得するように変更しますが、現時点ではこのユーザの投稿のみ取得します）
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+
+        // Welcomeビューでそれらを表示
+        return view('welcome',$data);
     }
+        
+    
 
     /**
      * Show the form for creating a new resource.
@@ -32,7 +43,7 @@ class TasksController extends Controller
     public function create()
     {
         //
-        $task = new Task;
+        $task = new Tasks;
         
         return view('tasks.create',['task'=> $task,]);
     }
@@ -51,12 +62,14 @@ class TasksController extends Controller
             'content' => 'required|max:255',
             ]);
             
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
         
-        return redirect('/');
+         // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
+        
+        header('Location:./');
     }
 
     /**
@@ -68,9 +81,11 @@ class TasksController extends Controller
     public function show($id)
     {
         //
-        $task = Task::findOrFail($id);
+        $task = Tasks::findOrFail($id);
         
-        return view('tasks.show',['task'=> $task,]);
+        return view('tasks.show',
+        ['task'=> $task,
+        ]);
     }
 
     /**
@@ -82,7 +97,7 @@ class TasksController extends Controller
     public function edit($id)
     {
         //
-        $task = Task::findOrFail($id);
+        $task = Tasks::findOrFail($id);
         
         return view('tasks.edit', ['task'=> $task,]);
     }
@@ -100,7 +115,7 @@ class TasksController extends Controller
         $request->validate([
             'status' => 'required|max:10',
             'content' => 'required|max:255',]);
-        $task = Task::findorFail($id);
+        $task = Tasks::findorFail($id);
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
@@ -119,9 +134,11 @@ class TasksController extends Controller
         //
         
         // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
+        $task = Tasks::findOrFail($id);
         // メッセージを削除
-        $task->delete();
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
